+++ resources/js/admin/entities/question/model/types.ts
@@ -1,4 +1,5 @@
 import {QuestionOption, QuestionOptionPayload} from "../../questionOption/model/types";
+import {QuestionAnswer, QuestionAnswerPayload} from "../../questionAnswer/model/types";
 
 export type QuestionType = "single_choice" | "multiple_choice" | "boolean" | "fill_in_blank" | "match" | "sentence_build" | "free_text";
 
@@ -11,10 +12,7 @@ export type Question = {
     type: QuestionType,
     settings?: {} | null,
     options: QuestionOption[]
-    answer: {
-        questionType: QuestionType,
-        value: string[],
-    },
+    answer: QuestionAnswer,
 };
 
 export type QuestionPayload = {
@@ -25,9 +23,6 @@ export type QuestionPayload = {
     points: number,
     orderIndex?: number | null,
     settings?: {} | null,
-    answer: {
-        questionType: QuestionType,
-        value: string[],
-    },
+    answer: QuestionAnswerPayload,
     options: QuestionOptionPayload[]
 };