+++ routes/react-admin.php
@@ -15,6 +15,7 @@ use App\UI\Http\Api\Admin\Controllers\Media\MediaFileUploadController;
 use App\UI\Http\Api\Admin\Controllers\Module\ModuleCreateController;
 use App\UI\Http\Api\Admin\Controllers\Module\ModuleShowController;
 use App\UI\Http\Api\Admin\Controllers\Module\ModuleUpdateController;
+use App\UI\Http\Api\Admin\Controllers\Question\QuestionAnswerPatchController;
 use App\UI\Http\Api\Admin\Controllers\QuestionsGroup\QuestionsGroupCreateController;
 use App\UI\Http\Api\Admin\Controllers\QuestionsGroup\QuestionsGroupDetailsController;
 use App\UI\Http\Api\Admin\Controllers\Quiz\QuizCreateController;
@@ -66,5 +67,9 @@ Route::middleware(['web'])->prefix('react/api')->group(function () {
     Route::prefix('questionGroups')->group(function () {
         Route::get('{uuid}', QuestionsGroupDetailsController::class);
         Route::post('/', QuestionsGroupCreateController::class);
+    });
+
+    Route::prefix('questions')->group(function () {
+        Route::patch('{uuid}/answer', QuestionAnswerPatchController::class);
     });
 });